Renamed src/consts.ts to src.class_consts.ts
