+++ src/class_consts.ts
