+++ src/css/common/vars/scope.scss
@@ -1,6 +1,6 @@
 /*
   All styles will only apply to elements, with a $pp-scope class parent.
   We do not use global styles, so they do not break other websites' styles.
-  IMPORTANT: keep it in sync with consts.ts
+  IMPORTANT: keep it in sync with class_consts.ts
  */
 $pp-scope: ".pp-ui";
