+++ src/css/selection.scss
@@ -1,6 +1,6 @@
 /*
   Styles related to selection highlighting
-  IMPORTANT: keep it in sync with consts.ts
+  IMPORTANT: keep it in sync with class_consts.ts
  */
 
 .pp-highlight {
